+++ CodeStarterCamp_Week1/iceCream.swift
@@ -1,0 +1,22 @@
+//
+//  step2.swift
+//  CodeStarterCamp_Week1
+//
+//  Created by 수염 on 2022/04/16.
+//
+
+import Foundation
+
+//MARK: 아이스크림 그리기
+func drawIceCream() {
+    for _ in 1...8 {
+        for _ in 1...11 {
+            print("*", terminator: "")
+        }
+    print()
+    }
+    for _ in 1...4 {
+        print("    | |")
+    }
+}
+
