+++ CodeStarterCamp_Week1/main.swift
@@ -9,18 +9,23 @@
 import Foundation
 
 
-//MARK: 아이스크림 그리기
-func drawIceCream() {
-    for _ in 1...8 {
-        for _ in 1...11 {
-            print("*", terminator: "")
-        }
-    print()
-    }
-    for _ in 1...4 {
-        print("    | |")
-    }
-}
 
+//MARK: 실행예시
+bodyLength = 10; stickLength = 4; choco = "***"; toping = " "
+makePepero()
 
-drawIceCream()
+bodyLength = 12; stickLength = 4; choco = "***"; toping = "&"
+makePepero()
+
+bodyLength = 12; stickLength = 6; choco = "***"; toping = "#"
+makePepero()
+
+bodyLength = 6; stickLength = 4; choco = "|0|"; toping = " "
+makePepero()
+
+//MARK: 내마음대로 만들기
+bodyLength = 10; stickLength = 5; choco = "##"; toping = "🔥"
+makePepero()
+
+bodyLength = 12; stickLength = 5; choco = "888"; toping = "☆"
+makePepero()
